+++ tests/PSRule.Tests/TestModule4/rules/Module4.Rule.ps1
@@ -7,8 +7,8 @@
 
 # Synopsis: Test rule in TestModule4
 Rule 'M4.Rule1' {
-    $Rule.TargetName -eq 'TestObject1'
-    $Rule.TargetType -eq 'TestObjectType'
+    $PSRule.TargetName -eq 'TestObject1'
+    $PSRule.TargetType -eq 'TestObjectType'
     $Configuration.ruleConfig1 -eq 'Test'
 }